+++ lib/dropdown_custom.dart
@@ -9,3 +9,4 @@ export 'src/custom_dropdown.dart'
     show CustomDropdown, ItemLabel, ItemGroup, ItemEnabled, DropdownItemBuilder;
 export 'src/dropdown_decoration.dart' show DropdownDecoration;
 export 'src/dropdown_direction.dart' show DropdownDirection;
+export 'src/dropdown_loading.dart' show DropdownLoading;